new _internal.h added to share definitions among sources; unnecessary and/or redundant things removed from sources; fixed several typos and other errors in README; improved temporary *put feedbacks in pr_Pointer methods; tuplet and mininstall updated

+++ include/prelude/order.h
@@ -74,7 +74,7 @@ pr_Order_sput_len(pr_Order self);
 /*- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - */
 /* This value represents the maximum buffer length required to store the
    qualified representation of order */
-#define PR_ORDER_MAX_SPUT_LEN ((size_t)33)
+#define PR_ORDER_MAX_SPUT_LEN 33
 
 /*- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - */
 /* Return : the opposite value of the passed order. For example, if pr_EQUAL
